double

+++ csrc/flashinfer_rope_ops.cu
@@ -42,7 +42,7 @@ void apply_rope_pos_ids_cos_sin_cache(at::Tensor q, at::Tensor k, at::Tensor q_r
 void mla_rope_quantize(at::Tensor q_rope_in, at::Tensor k_rope_in, at::Tensor q_nope_in,
                        at::Tensor k_nope_in, at::Tensor q_rope_out, at::Tensor k_rope_out,
                        at::Tensor q_nope_out, at::Tensor k_nope_out, at::Tensor pos_ids,
-                       float quant_scale_q, float quant_scale_kv, bool interleave);
+                       double quant_scale_q, double quant_scale_kv, bool interleave);
 
 TORCH_LIBRARY_FRAGMENT(TORCH_EXTENSION_NAME, m) {
   // "Apply RoPE"
